Move PreProcessor to project package
