+++ sit-cv-core/src/main/java/io/sitoolkit/cv/core/domain/project/PreProcessor.java
@@ -1,4 +1,4 @@
-package io.sitoolkit.cv.core.domain.project.preprocess;
+package io.sitoolkit.cv.core.domain.project;
 
 import java.nio.file.Path;
